Fixing more failing FlatObjectQueryBuilder tests
Providing more implementation for FlatObjectQueryBuilder so that
more tests from AbstractQueryTestcase are passing.

One particular issue is that several internal methods assume access to
package restricted members or methods. We provided workaround by
introducing TempWorkaround class that is a member for this
specific package and exposes required methods in a more broad manner.
This is a temporary solution and should go away once we decide whether
we need the QueryBuilder or not. And if we need it then we can move it
to specific package or consider different solution. However, we need
revisit this in the future.

Signed-off-by: Lukáš Vlček <lukas.vlcek@aiven.io>