Fix failing testExistsQueryMinimalMapping test
This test was failing because the builder instance that is passed into
multi-fields has the DocValues set to true by default.

By setting it to false made the test passing successfully. But this
needs to be investigated further.

Also fixed the yamlTest to pass successfully, however, it required to
call the internal `_value` field in the query explicitly. This can
mean that we will need to implement and employ the QueryBuilder(?).

Signed-off-by: Lukáš Vlček <lukas.vlcek@aiven.io>