Add Keyword Field and Refactor FieldMapper with ParameterizedFieldMapper (#3)
* Refactor FieldMapper to ParameterizedFieldMapper

Signed-off-by: Mingshi Liu <mingshl@amazon.com>

* Refactor Plugin with ParameterizedFieldMapper

Signed-off-by: Mingshi Liu <mingshl@amazon.com>

* Add Keyword Field and Tests
Signed-off-by: Mingshi Liu <mingshl@amazon.com>

* Update FlatObjectFieldMapperTests

Signed-off-by: Mingshi Liu <mingshl@amazon.com>

* Enable flat-object Plugin to run

Signed-off-by: Mingshi Liu <mingshl@amazon.com>

* Enable Search for each subfields

Signed-off-by: Mingshi Liu <mingshl@amazon.com>

* flat multiple nested fields

Signed-off-by: Mingshi Liu <mingshl@amazon.com>

* Fix some of the failing tests

1) Fixing failing test `FlatObjectFiledMapperTests.testExistsQueryMinimalMapping`:

```
Can not write a field name, expecting a value
com.fasterxml.jackson.core.JsonGenerationException: Can not write a field name, expecting a value
	at __randomizedtesting.SeedInfo.seed([163E47FC8A06B681:6F031AC0DD5FD2A1]:0)
	at app//com.fasterxml.jackson.core.JsonGenerator._reportError(JsonGenerator.java:2733)
	at app//com.fasterxml.jackson.core.json.UTF8JsonGenerator.writeFieldName(UTF8JsonGenerator.java:217)
	at app//org.opensearch.common.xcontent.json.JsonXContentGenerator.writeFieldName(JsonXContentGenerator.java:195)
	at app//org.opensearch.common.xcontent.XContentBuilder.field(XContentBuilder.java:298)
	at app//org.opensearch.flatobject.mapper.FlatObjectFieldMapperTests.writeFieldValue(FlatObjectFieldMapperTests.java:45)
	at app//org.opensearch.index.mapper.MapperTestCase.writeField(MapperTestCase.java:87)
	at app//org.opensearch.index.mapper.MapperServiceTestCase.source(MapperServiceTestCase.java:184)
	at app//org.opensearch.index.mapper.MapperTestCase.assertExistsQuery(MapperTestCase.java:108)
	at app//org.opensearch.index.mapper.MapperTestCase.testExistsQueryMinimalMapping(MapperTestCase.java:103)
```

The `writeFieldValue` needs to construct a real object.

2) Add "missing" GeoShape field mapper plugin for `FlatObjectQueryBuilderTests`:

This is really a hack, which is unfortunately required.
For further info see the following ticket I opened:
<https://github.com/opensearch-project/OpenSearch/issues/6322>

We are able to start fixing individual failing tests.

Signed-off-by: Lukáš Vlček <lukas.vlcek@aiven.io>

* parse path_content as string field

Signed-off-by: Mingshi Liu <mingshl@amazon.com>

* Fix yaml test folders

Yaml tests were placed under folder called `rest-api-spec.test` but this needs to be a read hierarchical folder structure `rest-api`/`test`. Also added one more subfolder.

Now we are able to run yaml tests and using `../../gradlew :plugins:flat-object:yamlRestTest` or directly from Idea.

Signed-off-by: Lukáš Vlček <lukas.vlcek@aiven.io>

* Fix the first yaml test

We have the first working yaml test!

Signed-off-by: Lukáš Vlček <lukas.vlcek@aiven.io>

* Fix some CI failures

- Add modification to CHANGELOG.md to pass CHANGELOG check
- Fix `javadoc` and `precommit` issues (mostly formatting and naming)

Signed-off-by: Lukáš Vlček <lukas.vlcek@aiven.io>

* Fix more CI failures

- Fixing testingConventions no.1

  Not extending OpenSearchTestCase (or respectively LuceneTestCase) was
  causing this error:

```
> Task :plugins:flat-object:testingConventions FAILED
Expected at least one test class included in task :plugins:flat-object:integTest, but found none.
Tests classes with suffix `Tests` should extend org.apache.lucene.tests.util.LuceneTestCase but the following classes do not:
  * org.opensearch.flatobject.xcontent.KeyValueJsonXContentParserTests
```
- Fixing testingConventions no.2

  Custom support of `gradle run` task was causing the following error:

```
> Task :plugins:flat-object:testingConventions FAILED
Expected at least one test class included in task :plugins:flat-object:integTest, but found none.

FAILURE: Build failed with an exception.

* What went wrong:
Execution failed for task ':plugins:flat-object:testingConventions'.
> Testing conventions [Expected at least one test class included in task :plugins:flat-object:integTest, but found none.] are not honored
```
  We fixed this by making part of `build.gradle` script conditional.
  This means that from now the `run` task needs to be called like this:

  `gradle -PisLocal run`

Signed-off-by: Lukáš Vlček <lukas.vlcek@aiven.io>

* Add FlatObjectMappingBenchmark

Signed-off-by: Mingshi Liu <mingshl@amazon.com>

* Fixing more failing FlatObjectQueryBuilder tests

Providing more implementation for FlatObjectQueryBuilder so that
more tests from AbstractQueryTestcase are passing.

One particular issue is that several internal methods assume access to
package restricted members or methods. We provided workaround by
introducing TempWorkaround class that is a member for this
specific package and exposes required methods in a more broad manner.
This is a temporary solution and should go away once we decide whether
we need the QueryBuilder or not. And if we need it then we can move it
to specific package or consider different solution. However, we need
revisit this in the future.

Signed-off-by: Lukáš Vlček <lukas.vlcek@aiven.io>

* Resolve merge conflict in FlatObjectFieldMapper.java

* Add subfields FieldMapper for -_value and ._valueAndPath

Signed-off-by: Mingshi Liu <mingshl@amazon.com>

* resolve CI test

Signed-off-by: Mingshi Liu <mingshl@amazon.com>

* Fix failing testExistsQueryMinimalMapping test

This test was failing because the builder instance that is passed into
multi-fields has the DocValues set to true by default.

By setting it to false made the test passing successfully. But this
needs to be investigated further.

Also fixed the yamlTest to pass successfully, however, it required to
call the internal `_value` field in the query explicitly. This can
mean that we will need to implement and employ the QueryBuilder(?).

Signed-off-by: Lukáš Vlček <lukas.vlcek@aiven.io>

* Fix spotless errors in benchmarks

Signed-off-by: Lukáš Vlček <lukas.vlcek@aiven.io>

* Improve index mapping test

Explicitly verify that no dynamic fields are created, hence the
flat-object field helps reduce the mapping explosion.

Signed-off-by: Lukáš Vlček <lukas.vlcek@aiven.io>

* add tests

Signed-off-by: Mingshi Liu <mingshl@amazon.com>

* Dynamic Create FlatObjectFieldType for Dotpath field

---------

Signed-off-by: Mingshi Liu <mingshl@amazon.com>
Signed-off-by: Lukáš Vlček <lukas.vlcek@aiven.io>
Co-authored-by: Lukáš Vlček <lukas.vlcek@aiven.io>